ENG-1282 - Fixing expression editor crashing perseus editor issue
This seems to be an old bug that just never showed up because we were not using the expression editor before. In this commit: https://github.com/LearnToBeFoundation/perseus/commit/d86625099945357c4c0c78c9e0e7de5d11f42d25 we see that all these instances of the Underscore library are changed to not wrap the object. In this conversion it looks like a mistake was made when dealing with this mapping over an object (there is no `.map` for objects, only for arrays). Underscore's `map` method *will* work for objects though, so we just put this code back to its original intent (using underscore's map method)