ENG-1282: Fixing further JS errors when using "all widgets" template
The "all widgets" testing template reveals 2 more JS errors -- one in the LikerT Scale and one in the Graphie component. The former was just missing some default params (it worked when you selected the item from the widget chooser but not in the "all widgets" template). The latter had an infinite recursion because it tries to do a deep comparison on previous props and next props -- however some of the elements are react components which contain circular references (particularly on the _owner field), so we just updated the deep compare to avoid that particular field on things that look like react components.